manager/app/requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==0.21.1
tailer==0.4.1
uWSGI==2.0.21
Werkzeug==2.2.2
zipp==3.13.0
